feat : useCommunity hook
