+++ frontend/recoil/atom.ts
@@ -9,9 +9,3 @@ export const checkState = atom({
   key: 'checkState',
   default: false,
 });
-
-// 데이터 필터링
-export const filterState = atom({
-  key: 'filterState',
-  default: 'sorted',
-});
